Vendor Ruby if an installation is not found
This buildpack requires that a Ruby installation be available during
buildpack compile, in order to install the `parseconfig` gem, and then
run the `install_git_submodules.rb` script.

On Heroku-22 the stack image no longer includes a system Ruby
installation, so in order for this buildpack to continue to work, a
Ruby install must be vendored by this buildpack.

If an app already uses the Ruby buildpack, and that buildpack is ordered
prior to the submodules buildpack, then this buildpack will skip the
Ruby vendoring step to save downloading a redundant copy of Ruby.

The Ruby installation vendored by this buildpack is not made available
to later buildpacks or at runtime, so will otherwise not affect the
build.

Fixes #7.